Minor code cleanup

+++ src/test/java/com/maxmind/minfraud/response/DeviceTest.java
@@ -23,7 +23,7 @@ public class DeviceTest extends AbstractOutputTest {
                         .finish()
         );
 
-        assertEquals(99.0, device.getConfidence().doubleValue(), 1e-15);
+        assertEquals(99.0, device.getConfidence(), 1e-15);
         assertEquals(UUID.fromString("C8D3BE1A-BE26-11E5-8C50-1B575C37265F"), device.getId());
         assertEquals("2016-06-08T14:16:38Z", device.getLastSeen());
     }
